Move tests to time.js
