+++ tests/test.js
@@ -59,7 +59,6 @@ describe('Endpoints', function() {
     require('./times')(expect, request, baseUrl);
     require('./activities')(expect, request, baseUrl);
     require('./projects')(expect, request, baseUrl);
-    require('./delete')(expect);
 
 });
